[mypy] Enable check for some files in nncf/common (#3171)
### Changes

1) Enable mypy check for:
    - nncf/api
    - nncf/common/*.py (exlude some files with many error)
    - nncf/common/initialization

2) Remove `extension_is_loading_info_log` - torch specific function that
used only ones.

+++ nncf/common/schedulers.py
@@ -69,7 +69,7 @@ class PolynomialDecaySchedule:
         else:
             value = self.initial_value + (self.target_value - self.initial_value) * np.power(progress, self.power)
 
-        return value
+        return float(value)
 
 
 class MultiStepSchedule:
@@ -141,7 +141,7 @@ class ExponentialDecaySchedule:
         if self.target_epoch == 0:
             return self.target_value
 
-        value = self.initial_value * np.power(self.decay_rate, epoch / self.target_epoch)
+        value = self.initial_value * float(np.power(self.decay_rate, epoch / self.target_epoch))
         return max(value, self.target_value)
